pyproject.toml
[project]
name = "tkinter-file-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter (>=5.2.2,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "blinker (>=1.9.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-benchmark = "^5.1.0"
